pyproject.toml
[project]
name = "lunchmoney-venmo-track"
version = "0.1.0"
description = ""
keywords = []
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "lunchable>=1.4.3",
    "python-decouple-typed>=3.11.0",
    "structlog-config>=0.10.0",
    "venmo-api-unofficial",
]
authors = [{ name = "Michael Bianco", email = "mike@mikebian.co" }]
urls = { "Repository" = "https://github.com/iloveitaly/lunchmoney-venmo-track" }

# additional packaging information: https://packaging.python.org/en/latest/specifications/core-metadata/#license
[project.scripts]
lunchmoney-venmo-track = "lunchmoney_venmo_track:main"

[build-system]
requires = ["uv_build>=0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
# avoids the src/ directory structure
module-root = ""

[tool.uv.sources]
venmo-api-unofficial = { git = "https://github.com/iloveitaly/venmo-api-unofficial", rev = "patch-1" }

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pyright[nodejs]>=1.1.408",
    "ruff>=0.15.0",
    "coverage>=7.13.4",
    "pytest-cov>=7.0.0",
    "covdefaults>=2.3.0",
]

[tool.pyright]
exclude = ["examples/", "playground/", "tmp/", ".venv/", "tests/"]

[tool.pytest.ini_options]
addopts = "--cov --cov-report=term-missing --cov-report=html:tmp/htmlcov"

[tool.coverage.run]
plugins = ["covdefaults"]
source = ["lunchmoney_venmo_track"]

[tool.coverage.report]
fail_under = 50

[tool.ruff]
extend-exclude = ["playground.py", "playground/"]
